requirements.txt
google-generativeai
google-genai
streamlit
python-dotenv
pymupdf
python-docx
